+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
-property injection tagged as @optional no longer overrides default values with `undefined`. 
+property injection tagged as @optional no longer overrides default values with `undefined`.
+Updated `targetName` to be a valid `typescript@5` decorator.
 
 ## [6.0.2]